requirements.txt
Flask==2.2.5
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.1
pandas==2.1.1
openpyxl==3.1.3

